feat(web): split trigger catalog page handlers [OPE-409]
